Refactor Google Integration and its utilities
